+++ src/argilla/server/apis/v1/handlers/datasets/datasets.py
@@ -30,14 +30,12 @@ from argilla.server.schemas.v1.datasets import (
     DatasetMetrics,
     Datasets,
     DatasetUpdate,
-    MetadataProperties,
-    MetadataProperty,
-    MetadataPropertyCreate,
     VectorSettings,
     VectorSettingsCreate,
     VectorsSettings,
 )
 from argilla.server.schemas.v1.fields import Field, FieldCreate, Fields
+from argilla.server.schemas.v1.metadata_properties import MetadataProperties, MetadataProperty, MetadataPropertyCreate
 from argilla.server.schemas.v1.questions import Question, QuestionCreate, Questions
 from argilla.server.search_engine import (
     SearchEngine,
